Cost: isolate the pricing fetch as the egress boundary
Structure-only commit: cost/remote.rs now holds everything that
touches the network in the cost pipeline — the LiteLLM URL, the fetch,
and the refresh entry points. Besides the opt-in Cursor collector this
is the binary's only egress, so a diff touching cost/remote.rs is an
egress change by definition; pricing math, state, and lookup stay in
cost.rs. refresh_pricing tightens to pub(super) (callers only ever
used spawn_pricing_refresh, whose path is unchanged).

Second-model review round 1 clean; same-instant A/B against a HEAD
worktree build renders byte-identical.
